swastik-hospital/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pymongo>=4.6.0
motor>=3.3.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
